[FEAT]: pin 클릭시 해당 위치의 메뉴 불러오기 구현

+++ app/src/main/java/com/kuit/ourmenu/data/model/map/response/MapDetailResponse.kt
@@ -20,11 +20,11 @@ data class MapDetailResponse(
     @SerialName("menuFolderInfo")
     val menuFolderInfo: MenuFolderInfo,
     @SerialName("mapId")
-    val mapId: Int,
+    val mapId: Long,
     @SerialName("mapX")
-    val mapX: Int,
+    val mapX: Double,
     @SerialName("mapY")
-    val mapY: Int
+    val mapY: Double
 )
 
 @Serializable
